fix: router reactjs and add model
